+++ src/Novu.php
@@ -12,6 +12,8 @@ class Novu
         Actions\ManagesTopics,
         Actions\ManagesSubscribers,
         Actions\ManagesActivity,
+        Actions\ManagesInboundParse,
+        Actions\ManagesNotifications,
         Actions\ManagesChanges,
         Actions\ManagesEnvironments,
         Actions\ManagesExecutionDetails,