+++ CHANGELOG.md
@@ -1,8 +1,35 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/)
-and this project adheres to [Semantic Versioning](http://semver.org/).
+1.0.5 / 2023-04-06
+===================
 
-## [Unreleased]
-- Adds first version
+* Add support for retrieving a notification, all notifications, notification stats, and notification graph stats.
+* Add support for validating mx record for inbound parse functionality 
+
+1.0.4 / 2023-03-03
+===================
+
+* Add support for triggering bulk events. 
+
+1.0.3 / 2023-02-17
+===================
+
+* Add support for Topics. 
+
+1.0.2 / 2023-01-10
+===================
+
+* Psr-4 support fix for IsNull.php. 
+
+1.0.1 / 2023-01-10
+==================
+
+* Bug Fixes.
+  
+1.0.0 / 2023-01-09
+==================
+
+* Initial Release.
+